requirements.txt
# Core dependencies
pyyaml>=6.0
jsonschema>=4.17.0
requests>=2.28.0
pygithub>=2.1.0

# Python analyzers
ruff>=0.1.0
pylint>=3.0.0
bandit>=1.7.5
mypy>=1.7.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
responses>=0.24.0
